Refactor complete on About page
